+++ igrep/src/indexer/Index.cpp
@@ -7,7 +7,7 @@
 #include<filesystem>
 #include<ios>
 #include<unordered_set>
-
+#include<cstdint>
 
 using namespace std;
 using namespace std::filesystem;
@@ -17,7 +17,7 @@ using namespace igrep::utils;
 namespace igrep::indexer {
 
 	
-	void Index::process_line(const string& line, const string& filename, const size_t& line_number, size_t& word_index) {
+	void Index::process_line(const string& line, const string& filename, const uint32_t& line_number, uint32_t& word_index) {
 		path absolute_path = path(absolute(filename));
 		absolute_path = canonical(absolute_path);
 		indexed_files.insert(absolute_path);
@@ -26,8 +26,8 @@ namespace igrep::indexer {
 		string normalized =  StringUtils::normalize_line(line);
 		istringstream iss(normalized);
 		string word;
-		size_t indent;
-		size_t start_pos = 0;
+		uint32_t indent;
+		uint32_t start_pos = 0;
 		while (iss >> word) {
 			indent = lower_case_line.find(word, start_pos);
 			start_pos = indent + 1;
@@ -85,30 +85,30 @@ namespace igrep::indexer {
 		if (!ofs.is_open()) {
 			throw runtime_error("Cannot open file " + path);
 		}
-		size_t set_size = indexed_files.size();
+		uint32_t set_size = indexed_files.size();
 		ofs.write(reinterpret_cast<const char*>(&set_size), sizeof(set_size));
 		for(const auto& el : indexed_files){
 			string string_path = el.string();
-			size_t path_size = string_path.length();
+			uint32_t path_size = string_path.length();
 			ofs.write(reinterpret_cast<const char*>(&path_size), sizeof(path_size));
 			ofs.write(string_path.data(), path_size);
 		}
 
-		size_t map_size = words.size();
+		uint32_t map_size = words.size();
 		ofs.write(reinterpret_cast<char*>(&map_size), sizeof(map_size));
 
 		for (const auto& pair : words) {
 
-			size_t key_length = pair.first.length();
+			uint32_t key_length = pair.first.length();
 			ofs.write(reinterpret_cast<const char*>(&key_length), sizeof(key_length));
 			ofs.write(pair.first.data(), key_length);
 
-			size_t vector_size = pair.second.size();
+			uint32_t vector_size = pair.second.size();
 			ofs.write(reinterpret_cast<const char*>(&vector_size), sizeof(vector_size));
 
 			for (const Position& position : pair.second) {
 
-				size_t filename_length = position.filename.length();
+				uint32_t filename_length = position.filename.length();
 				ofs.write(reinterpret_cast<const char*>(&filename_length), sizeof(filename_length));
 				ofs.write(position.filename.data(), filename_length);
 
@@ -126,11 +126,11 @@ namespace igrep::indexer {
 			if (!ifs.is_open()) {
 				throw runtime_error("Cannot open index file" + path);
 			}
-			size_t set_size;
+			uint32_t set_size;
 			ifs.read(reinterpret_cast<char*>(&set_size), sizeof(set_size));
 
 			for (int i = 0; i < set_size; i++){
-				size_t path_size;
+				uint32_t path_size;
 				ifs.read(reinterpret_cast<char*>(&path_size), sizeof(path_size));
 				string file_string;
 				file_string.resize(path_size);
@@ -142,34 +142,34 @@ namespace igrep::indexer {
 
 
 
-			size_t map_size;
+			uint32_t map_size;
 			ifs.read(reinterpret_cast<char*>(&map_size), sizeof(map_size));
 
 			for (int i = 0; i < map_size; i++) {
-				size_t key_len;
+				uint32_t key_len;
 				ifs.read(reinterpret_cast<char*>(&key_len), sizeof(key_len));
 
 				string key;
 				key.resize(key_len);
 				ifs.read(key.data(), key_len);
 
-				size_t vector_size;
+				uint32_t vector_size;
 				ifs.read(reinterpret_cast<char*>(&vector_size), sizeof(vector_size));
 
 				vector<Position> positions;
 				positions.reserve(vector_size);
 
 				for (int j = 0; j < vector_size; j++) {
-					size_t filename_len;
+					uint32_t filename_len;
 					ifs.read(reinterpret_cast<char*>(&filename_len), sizeof(filename_len));
 
 					string filename;
 					filename.resize(filename_len);
 					ifs.read(filename.data(), filename_len);
 
-					size_t indent;
-					size_t line_number;
-					size_t word_index;
+					uint32_t indent;
+					uint32_t line_number;
+					uint32_t word_index;
 
 					ifs.read(reinterpret_cast<char*>(&word_index), sizeof(word_index));
 					ifs.read(reinterpret_cast<char*>(&indent), sizeof(indent));
